Update db connection for testing
