+++ database/seed.js
@@ -1,45 +1,6 @@
 /* eslint-disable no-console */
-const mysql = require('mysql2/promise');
 const faker = require('faker/locale/en_US');
-const fs = require('fs');
-const path = require('path');
-const auth = require('./auth');
-
-const createDbConn = (scopeAuth, env) => {
-  const {
-    user, password, host,
-  } = process.env.NODE_ENV === 'test' ? scopeAuth.authTest : scopeAuth.authTest;
-  const database = env === 'test' ? 'fRiend-test' : 'fRiend';
-
-  return mysql.createConnection({
-    host,
-    user,
-    password,
-    database,
-    multipleStatements: true,
-  });
-};
-
-const createDbTables = (conn) => {
-  const schemaFile = path.resolve(__dirname, 'schema.sql');
-  const createDBQuery = fs.readFileSync(schemaFile).toString();
-
-  return conn.query(createDBQuery);
-};
-
-const cleanDbTables = (conn) => {
-  const query = `
-    SET FOREIGN_KEY_CHECKS = 0;
-
-    TRUNCATE TABLE rates;
-    TRUNCATE TABLE lenders;
-    TRUNCATE TABLE properties;
-    TRUNCATE TABLE zips;
-
-    SET FOREIGN_KEY_CHECKS = 1;
-  `;
-  return conn.query(query);
-};
+const { dbConn, createDbTables, cleanDbTables } = require('./index');
 
 const seedZips = (conn, zips) => {
   const taxLow = 0.8;
@@ -166,7 +127,7 @@ const seedRates = (conn, zips) => {
   return conn.query(query);
 };
 
-const seedDb = async (scopeAuth, env) => {
+const seedDb = async (conn, env) => {
   let sharedZips = new Set();
   while (sharedZips.size < 10) {
     const zip = faker.address.zipCode();
@@ -175,8 +136,6 @@ const seedDb = async (scopeAuth, env) => {
     }
   }
   sharedZips = [...sharedZips];
-
-  const conn = await createDbConn(scopeAuth);
 
   await createDbTables(conn, env);
   console.log('created database tables if non-existant');
@@ -196,9 +155,7 @@ const seedDb = async (scopeAuth, env) => {
   await seedRates(conn, sharedZips);
   console.log('seeded rates table');
 
-  conn.close();
+  conn.end();
 };
 
-seedDb(auth).catch(console.log);
-
-module.exports = seedDb;
+seedDb(dbConn.promise()).catch(console.log);
